2 Helper functions for posterior evaluation
These 2 files are used as helper functions for evaluating the posterior distribution of ECM and EIS data. The "evaluate_posterior_distribution" defines functions that scores MCMC posterior distributions of each component in the ECM. It gives a functionality score for each component based on R^2 from a QQ-plot to a normal distribution, skewness of the sample and the number of peaks appeared on the distribution.

The script "update_circuits_info" complements "evaluate_posterior_distribution" by applying the posterior quality evaluation to a list of AutoEIS inference results and updating a circuits_df DataFrame accordingly.